requirement.txt
ansicon==1.89.0
blessed==1.20.0
click==8.1.7
colorama==0.4.6
editor==1.6.6
inquirer==3.2.4
jinxed==1.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
pydobot @ git+https://github.com/luismesas/pydobot.git@2ff8d7e1318ac9b7bc32bd33ff327d343f44927c
Pygments==2.17.2
pyserial==3.4
readchar==4.0.5
rich==13.7.0
runs==1.2.2
shellingham==1.5.4
six==1.16.0
termcolor==2.4.0
typer==0.9.0
typing_extensions==4.9.0
wcwidth==0.2.13
xmod==1.8.1
yaspin==3.0.1
